+++ packages/@cdklabs/genai-idp-bedrock-llm-processor/src/internal/classification-function.ts
@@ -4,6 +4,7 @@ SPDX-License-Identifier: Apache-2.0
 */
 
 import path from "path";
+import * as bedrock from "@aws-cdk/aws-bedrock-alpha/bedrock";
 import { PythonFunction } from "@aws-cdk/aws-lambda-python-alpha";
 import {
   IdpPythonFunctionOptions,
@@ -12,7 +13,6 @@ import {
   ITrackingTable,
   LogLevel,
 } from "@cdklabs/genai-idp";
-import * as bedrock from "@cdklabs/generative-ai-cdk-constructs/lib/cdk-lib/bedrock";
 import { Duration, Stack } from "aws-cdk-lib";
 import { Metric } from "aws-cdk-lib/aws-cloudwatch";
 import { ITable } from "aws-cdk-lib/aws-dynamodb";
@@ -85,7 +85,7 @@ export interface ClassificationFunctionProps extends IdpPythonFunctionOptions {
    * The Bedrock model to use for classification.
    * The AI model that will classify document pages.
    */
-  readonly classificationModel: bedrock.IInvokable;
+  readonly classificationModel: bedrock.IBedrockInvokable;
 
   /**
    * Optional Bedrock guardrail to apply to classification model interactions.